MH3-114735: remove vision applications
The current issue is that vision applications are part of the vision repository. When a new version
is released or tagged in the vision repository, the vision applications must be regenerated to update
version dependencies, and a new commit must be added to an already released version. To avoid these
version conflicts, moving the vision applications to a separate repository.

https://bitbucket.microchip.com/scm/mg/vision_apps.git

Signed-off-by: Sandeep Sheriker Mallikarjun <sandeep.sheriker@microchip.com>